requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
tavily-python
python-dotenv
textblob
tiktoken